Extract program lowering traversal
